refactor(links): move prosemirror plugin into own file
